Upgrade to bitter 0.5

+++ src/network/mod.rs
@@ -185,7 +185,7 @@ pub(crate) fn parse<'a>(
         .map(|(obj_id, attrs)| {
             let id = *obj_id;
             let max = attrs.keys().map(|&x| i32::from(x)).max().unwrap_or(2) + 1;
-            let max_bit_width = bitter::bit_width(max as u32);
+            let max_bit_width = bitter::bit_width(max as u64);
             Ok((
                 id,
                 CacheInfo {
@@ -201,7 +201,7 @@ pub(crate) fn parse<'a>(
 
     // 1023 stolen from rattletrap
     let max_channels = header.max_channels().unwrap_or(1023) as u32;
-    let channel_width = bitter::bit_width(max_channels) as i32 - 1;
+    let channel_width = bitter::bit_width(u64::from(max_channels)) as i32 - 1;
     let channel_bits = cmp::max(channel_width, 0);
     let num_frames = header.num_frames();
     let is_lan = header.match_type().map(|x| x == "Lan").unwrap_or(false);
